Apply patch 9.

+++ fractint/headers/fractint.h
@@ -513,7 +513,7 @@ struct fractalspecificstuff
    int (*per_pixel)(void);      /* once-per-pixel init */
    int (*per_image)(void);      /* once-per-image setup */
    int (*calctype)(void);       /* name of main fractal function */
-   int orbit_bailout;           /* usual bailout value for orbit calc */
+   long orbit_bailout;          /* usual bailout value for orbit calc */
 };
 
 struct alternatemathstuff
